webapp_tradingview_style/requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
requests
python-dotenv
lightweight-charts
plotly
dash
dash-bootstrap-components
streamlit
bokeh
